requirements.txt
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
openai==1.54.3
pydantic==2.9.2